+++ src/app/(common)/(window)/(pokemon)/(part)/partOne.tsx
@@ -1,14 +1,18 @@
+"use client";
+
 import { PART_ONE_DATA } from "@/constant/constant";
 import useTyping from "@/hooks/useTyping";
 import { ChevronDoubleDownIcon } from "@heroicons/react/16/solid";
 import Image from "next/image";
 import { useEffect } from "react";
 
+/* 2024/05/05 - loading 끝난 후 보여질 애니메이션 (story)*/
 export default function PartOne({ setStep }: { setStep: () => void }) {
   const { showTxt, displayedTxt, handleBoard, fade } = useTyping({
     txt: PART_ONE_DATA,
   });
 
+  /* 다음 대사 없을 시 페이지 전환을 위한 setStep */
   useEffect(() => {
     if (fade) setStep();
   }, [fade]);
@@ -22,6 +26,7 @@ export default function PartOne({ setStep }: { setStep: () => void }) {
         height={150}
         style={{ width: "auto", height: "auto" }}
       />
+      {/* 대사창 */}
       <div
         className="absolute w-full h-1/3 left-0 bottom-0 flex flex-col justify-end border-2 border-solid border-black p-[5px] bg-white transition-custom"
         onClick={handleBoard}
